refactor: move db write, read code to postprocess.py
